python/requirements.txt
pyserial
requests
websocket-client
python-socketio[client]